Move envelope types into separate submodule
